+++ blocks/userdetails/shareratio.php
@@ -10,7 +10,7 @@ if ($user['paranoia'] < 2 || $CURUSER['id'] === $user['id'] || $CURUSER['class']
             <td class="rowhead" style="vertical-align: middle">' . $lang['userdetails_share_ratio'] . '</td>
             <td>
                 <div class="level-left">
-                    ' . member_ratio((int) $user['uploaded'], (int) $user['downloaded']) . '
+                    ' . member_ratio($user['uploaded'], $user['downloaded']) . '
                     <span class="left10">' . get_user_ratio_image($user['uploaded'], $user['downloaded']) . '</span>
                 </div>
             </td>