+++ include/function_returnto.php
@@ -9,8 +9,8 @@ use Rakit\Validation\Validator;
 /**
  * @param string $partial_url
  *
- * @throws DependencyException
  * @throws NotFoundException
+ * @throws DependencyException
  *
  * @return bool|mixed
  */