+++ unused/offers.php.php
@@ -299,7 +299,7 @@ switch ($action) {
   <tr>
   <td>offered by:</td>
   <td>' . format_username((int) $usersdata['id']) . ' [ ' . get_user_class_name((int) $usersdata['class']) . ' ]
-  ratio: ' . member_ratio((int) $usersdata['uploaded'], (int) $usersdata['downloaded']) . get_user_ratio_image($usersdata['uploaded'], ($usersdata['downloaded'])) . '</td>
+  ratio: ' . member_ratio($usersdata['uploaded'], $usersdata['downloaded']) . get_user_ratio_image($usersdata['uploaded'], ($usersdata['downloaded'])) . '</td>
   </tr>' . ($arr['filled_torrent_id'] > 0 ? '<tr>
   <td>filled:</td>
   <td><a class="is-link" href="details.php?id=' . $arr['filled_torrent_id'] . '">yes, click to view torrent!</a></td>